Updated logic to prevent the NPE from happening in time trail. Moved the participants to there own package. Added a new key for the participant logic to prevent duplicate records in time trail. Updated logic to prevent an NPE from happening due to this new key being used to determine if we should add the participant object to the list.
