+++ src/main/java/f1/data/SessionInitializationResult.java
@@ -2,7 +2,7 @@ package f1.data;
 
 import f1.data.enums.FormulaEnum;
 import f1.data.mapKeys.DriverPair;
-import f1.data.parse.packets.ParticipantData;
+import f1.data.parse.packets.participant.ParticipantData;
 import f1.data.parse.packets.session.SessionData;
 
 import java.util.List;
