+++ src/main/java/f1/data/parse/packets/participant/ParticipantKey.java
@@ -1,0 +1,4 @@
+package f1.data.parse.packets.participant;
+
+public record ParticipantKey(int driverId, int raceNumber, int networkId, String lastName) {
+}
