+++ src/main/java/f1/data/ui/panels/dashboards/SetupInfoDashboard.java
@@ -1,29 +1,29 @@
 package f1.data.ui.panels.dashboards;
 
+import f1.data.enums.VisualTireEnum;
 import f1.data.parse.individualLap.CarSetupInfo;
-import f1.data.parse.individualLap.CarStatusInfo;
 import javafx.scene.control.Label;
 import javafx.scene.control.TitledPane;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.VBox;
-import f1.data.parse.packets.CarSetupData;
-import f1.data.enums.VisualTireEnum;
 
 
 public class SetupInfoDashboard extends VBox {
 
     public SetupInfoDashboard(String setupName, CarSetupInfo setupInfo, int visualCompound) {
         GridPane setupDetails = new GridPane();
-        //Prints the data out in the order that the setup elements exist in the setup menu in the game.
-        for (int i = 0; i < setupInfo.getSetupDashboardData().length; i++) {
-            Label[] inner = setupInfo.getSetupDashboardData()[i];
-            for (int j = 0; j < inner.length; j++) {
-                setupDetails.add(inner[j], i, j);
+        if (setupInfo != null) {
+            //Prints the data out in the order that the setup elements exist in the setup menu in the game.
+            for (int i = 0; i < setupInfo.getSetupDashboardData().length; i++) {
+                Label[] inner = setupInfo.getSetupDashboardData()[i];
+                for (int j = 0; j < inner.length; j++) {
+                    setupDetails.add(inner[j], i, j);
+                }
             }
+            //Prints the tire compound and its label out below the Fuel Load.
+            setupDetails.add(new Label("Tire "), 0, 1);
+            setupDetails.add(new Label(VisualTireEnum.fromValue(visualCompound).getDisplay()), 1, 1);
         }
-        //Prints the tire compound and its label out below the Fuel Load.
-        setupDetails.add(new Label("Tire "), 0, 1);
-        setupDetails.add(new Label(VisualTireEnum.fromValue(visualCompound).getDisplay()), 1, 1);
 
         this.setupName = new TitledPane(setupName, setupDetails);
